chore: fix some circular dependency issues.
