+++ nemo/collections/llm/gpt/data/mock.py
@@ -68,7 +68,7 @@ class MockDataModule(pl.LightningDataModule):
         vocab_file: Optional[str] = None,
         merges_file: Optional[str] = None,
     ):
-        from nemo.lightning.pytorch.callbacks.callback_group import CallbackGroup
+        from nemo.lightning.callback_group import CallbackGroup
 
         CallbackGroup.get_instance().on_dataloader_init_start()
