+++ nemo/collections/llm/t5/data/mock.py
@@ -49,7 +49,7 @@ class MockDataModule(pl.LightningDataModule):
         persistent_workers: bool = False,
         create_attention_mask: bool = False,
     ):
-        from nemo.lightning.pytorch.callbacks.callback_group import CallbackGroup
+        from nemo.lightning.callback_group import CallbackGroup
 
         CallbackGroup.get_instance().on_dataloader_init_start()
