+++ nemo/lightning/pytorch/callbacks/model_checkpoint.py
@@ -29,7 +29,7 @@ from lightning.pytorch.utilities import rank_zero_info
 
 from nemo.lightning.ckpt_utils import ckpt_to_dir
 from nemo.lightning.io.pl import TrainerContext
-from nemo.lightning.pytorch.callbacks.callback_group import CallbackGroup
+from nemo.lightning.callback_group import CallbackGroup
 from nemo.utils import logging
 from nemo.utils.app_state import AppState
